refactor: rename import directory to match package name
